Move TCP connectivity test functions to new module
Moved RunConnectivityTest, RunNoConnectivityTest, et al from dataplane
to a new "tcp" module/package. This separates the actual connectivity code
from the dataplane Ginko specs so one can use the connectivity test code
from other specs w/o also causing the dataplane specs to run.
